+++ packages/headless/doc-parser/doc-parser.ts
@@ -11,13 +11,19 @@ const entryPoint = apiPackage.entryPoints[0];
 
 const controllers: ControllerConfiguration[] = [
   {
+    initializer: 'buildHistoryManager',
+    samplePaths: {},
+  },
+  {
     initializer: 'buildFacet',
     samplePaths: {
       react_class: [
         'packages/samples/headless-react/src/components/facet/facet.class.tsx',
+        'packages/samples/headless-react/src/components/facet/facet-search.tsx',
       ],
       react_fn: [
         'packages/samples/headless-react/src/components/facet/facet.fn.tsx',
+        'packages/samples/headless-react/src/components/facet/facet-search.tsx',
       ],
     },
   },
@@ -35,7 +41,20 @@ const controllers: ControllerConfiguration[] = [
   },
   {
     initializer: 'buildNumericFacet',
+    samplePaths: {
+      react_class: [
+        'packages/samples/headless-react/src/components/numeric-facet/numeric-facet.class.tsx',
+      ],
+      react_fn: [
+        'packages/samples/headless-react/src/components/numeric-facet/numeric-facet.fn.tsx',
+      ],
+    },
+    utils: ['buildNumericRange'],
+  },
+  {
+    initializer: 'buildDateFacet',
     samplePaths: {},
+    utils: ['buildDateRange'],
   },
   {
     initializer: 'buildPager',
@@ -58,7 +77,16 @@ const controllers: ControllerConfiguration[] = [
   },
   {
     initializer: 'buildCategoryFacet',
-    samplePaths: {},
+    samplePaths: {
+      react_class: [
+        'packages/samples/headless-react/src/components/category-facet/category-facet.class.tsx',
+        'packages/samples/headless-react/src/components/category-facet/category-facet-search.tsx',
+      ],
+      react_fn: [
+        'packages/samples/headless-react/src/components/category-facet/category-facet.fn.tsx',
+        'packages/samples/headless-react/src/components/category-facet/category-facet-search.tsx',
+      ],
+    },
   },
   {
     initializer: 'buildDidYouMean',
@@ -92,6 +120,10 @@ const controllers: ControllerConfiguration[] = [
         'packages/samples/headless-react/src/components/query-error/query-error.fn.tsx',
       ],
     },
+  },
+  {
+    initializer: 'buildBreadcrumbManager',
+    samplePaths: {},
   },
   {
     initializer: 'buildRecommendationList',
@@ -150,6 +182,10 @@ const controllers: ControllerConfiguration[] = [
     },
   },
   {
+    initializer: 'buildCartRecommendationsList',
+    samplePaths: {},
+  },
+  {
     initializer: 'buildSearchBox',
     samplePaths: {
       react_class: [
@@ -181,6 +217,17 @@ const controllers: ControllerConfiguration[] = [
       ],
       react_fn: [
         'packages/samples/headless-react/src/components/results-per-page/results-per-page.fn.tsx',
+      ],
+    },
+  },
+  {
+    initializer: 'buildRelevanceInspector',
+    samplePaths: {
+      react_class: [
+        'packages/samples/headless-react/src/components/relevance-inspector/relevance-inspector.class.tsx',
+      ],
+      react_fn: [
+        'packages/samples/headless-react/src/components/relevance-inspector/relevance-inspector.fn.tsx',
       ],
     },
   },