+++ packages/headless/doc-parser/src/interface-resolver.ts
@@ -107,7 +107,13 @@ function resolvePropertySignature(
   ancestorNames: string[]
 ) {
   const typeExcerpt = p.propertyTypeExcerpt.spannedTokens[0];
+  const typeName = extractTypeName(p.propertyTypeExcerpt);
+
   if (isRecordType(typeExcerpt)) {
+    return buildEntityFromProperty(p);
+  }
+
+  if (isUnionType(typeName)) {
     return buildEntityFromProperty(p);
   }
 
@@ -189,11 +195,7 @@ function resolveMethodSignature(
   const params = m.parameters.map((p) =>
     resolveParameter(entry, p, ancestorNames)
   );
-  const typeExcerpt = m.returnTypeExcerpt.spannedTokens[0];
-  let returnType: AnyEntity = buildReturnTypeEntity(m);
-  if (isReference(typeExcerpt)) {
-    returnType = buildObjEntityFromReturnType(entry, m, ancestorNames);
-  }
+  const returnType = resolveMethodReturnType(entry, m, ancestorNames);
 
   return buildFuncEntity({
     name: m.displayName,
@@ -201,6 +203,24 @@ function resolveMethodSignature(
     params,
     returnType,
   });
+}
+
+function resolveMethodReturnType(
+  entry: ApiEntryPoint,
+  m: ApiMethodSignature,
+  ancestorNames: string[]
+) {
+  const typeExcerpt = m.returnTypeExcerpt.spannedTokens[0];
+
+  if (isPromise(typeExcerpt)) {
+    return buildReturnTypeEntity(m);
+  }
+
+  if (isReference(typeExcerpt)) {
+    return buildObjEntityFromReturnType(entry, m, ancestorNames);
+  }
+
+  return buildReturnTypeEntity(m);
 }
 
 function buildObjEntityFromReturnType(
@@ -223,9 +243,14 @@ export function resolveParameter(
   ancestorNames: string[]
 ) {
   const typeExcerpt = p.parameterTypeExcerpt.spannedTokens[0];
+  const type = p.parameterTypeExcerpt.text;
 
   if (isTypeAlias(typeExcerpt)) {
     return buildEntityFromParamAndResolveTypeAlias(entry, p);
+  }
+
+  if (isUnionType(type)) {
+    return buildParamEntity(p);
   }
 
   if (isReference(typeExcerpt)) {
@@ -296,6 +321,15 @@ function isRecordType(token: ExcerptToken) {
   return isReference(token) && isRecord;
 }
 
+function isPromise(token: ExcerptToken) {
+  const isPromise = token.text === 'Promise';
+  return isReference(token) && isPromise;
+}
+
 function isReference(token: ExcerptToken) {
   return token.kind === ExcerptTokenKind.Reference;
+}
+
+function isUnionType(typeName: string) {
+  return typeName.includes('|');
 }