+++ packages/headless/src/api/search/search/user-identity.ts
@@ -1,5 +1,0 @@
-export interface UserIdentity {
-  name: string;
-  provider: string;
-  type: string;
-}